refactor: astro-md -> legacy-astro-flavored-md, astro-md-mode -> astro-markdown
