--- packages/astro/test/fixtures/astro-markdown-md-mode/src/pages/code-in-md.md
@@ -1,7 +1,0 @@
-# Fenced code blocks
-
-```html
-<body>
-	<div>This should also work without any problems.</div>
-</body>
-```
