+++ packages/astro/test/fixtures/astro-markdown/astro.config.mjs
@@ -1,12 +1,8 @@
 import { defineConfig } from 'astro/config';
-import preact from '@astrojs/preact';
 import svelte from "@astrojs/svelte";
 
 // https://astro.build/config
 export default defineConfig({
-  integrations: [preact(), svelte()],
+  integrations: [svelte()],
   site: 'https://astro.build/',
-	legacy: {
-		astroFlavoredMarkdown: true,
-	}
 });
