+++ packages/astro/test/fixtures/astro-markdown/package.json
@@ -1,9 +1,8 @@
 {
-  "name": "@test/astro-markdown",
+  "name": "@test/astro-markdown-md-mode",
   "version": "0.0.0",
   "private": true,
   "dependencies": {
-    "@astrojs/preact": "workspace:*",
     "@astrojs/svelte": "workspace:*",
     "astro": "workspace:*"
   }
