+++ packages/astro/test/fixtures/astro-markdown/src/components/Counter.svelte
