+++ packages/astro/test/fixtures/astro-markdown/src/pages/code-in-md.md
@@ -1,12 +1,3 @@
-# Inline code blocks
-
-`<script>` tags in **Astro** components are now built,
-bundled and optimized by default.
-
-> Markdown formatting still works between tags in inline code blocks.
-
-We can also use closing `</script>` tags without any problems.
-
 # Fenced code blocks
 
 ```html
