+++ packages/astro/test/fixtures/legacy-astro-flavored-markdown/astro.config.mjs
@@ -1,8 +1,12 @@
 import { defineConfig } from 'astro/config';
+import preact from '@astrojs/preact';
 import svelte from "@astrojs/svelte";
 
 // https://astro.build/config
 export default defineConfig({
-  integrations: [svelte()],
+  integrations: [preact(), svelte()],
   site: 'https://astro.build/',
+	legacy: {
+		astroFlavoredMarkdown: true,
+	}
 });
