+++ packages/astro/test/fixtures/legacy-astro-flavored-markdown/src/components/SvelteButton.svelte
