+++ packages/astro/test/fixtures/legacy-astro-flavored-markdown/src/layouts/content.astro
