+++ packages/astro/test/fixtures/legacy-astro-flavored-markdown/src/pages/empty-code.md
