+++ packages/astro/test/fixtures/legacy-astro-flavored-markdown/src/pages/imported-md/with-components.astro
