+++ packages/astro/test/fixtures/legacy-astro-flavored-markdown/src/pages/namespace.md
