+++ packages/astro/test/fixtures/legacy-astro-flavored-markdown/src/pages/raw-content.json.js
@@ -1,4 +1,4 @@
-import { rawContent, compiledContent } from './basic.md';
+import { rawContent, compiledContent } from '../imported-md/with-components.md';
 
 export async function get() {
 	return {
