+++ packages/astro/test/fixtures/legacy-astro-flavored-markdown/src/pages/vite-env-vars.md
