+++ packages/astro/test/fixtures/legacy-astro-flavored-markdown/src/scripts/test.js
